prototype/requirements.txt
PyQt5==5.15.2
mysql-connector-python==8.0.22